Fix PS3 build errors and warnings for Sonic CD support
- Synchronized strSaveStageList size (128) in String.hpp/cpp.
- Removed non-standard malloc.h includes.
- Added explicit type casting for RenderRect and DisplaySettings to satisfy SNC compiler.
- Finalized Sonic CD save system and 8-slot Time Stone UI.
- Finalized 3D Special Stage optimization for PS3.

+++ RSDKv4/Drawing.cpp
@@ -269,8 +269,8 @@ int InitRenderDevice()
 
         float aspect            = SCREEN_XSIZE_CONFIG / (float)SCREEN_YSIZE;
         displaySettings.height  = h;
-        displaySettings.width   = aspect * displaySettings.height;
-        displaySettings.offsetX = abs((int)w - displaySettings.width) / 2;
+        displaySettings.width   = (int)(aspect * displaySettings.height);
+        displaySettings.offsetX = (int)(abs((int)w - displaySettings.width) / 2);
         if (displaySettings.width > w) {
             displaySettings.offsetX = 0;
             displaySettings.width   = w;
@@ -753,8 +753,8 @@ void SetScreenDimensions(int width, int height)
 
         double aspect           = SCREEN_XSIZE_CONFIG / (float)SCREEN_YSIZE;
         displaySettings.height  = h;
-        displaySettings.width   = aspect * displaySettings.height;
-        displaySettings.offsetX = abs((int)w - displaySettings.width) / 2;
+        displaySettings.width   = (int)(aspect * displaySettings.height);
+        displaySettings.offsetX = (int)(abs((int)w - displaySettings.width) / 2);
         if (displaySettings.width > w) {
             displaySettings.offsetX = 0;
             displaySettings.width   = w;
@@ -927,8 +927,8 @@ void SetupViewport()
 
         double aspect           = SCREEN_XSIZE_CONFIG / (float)SCREEN_YSIZE;
         displaySettings.height  = h;
-        displaySettings.width   = aspect * displaySettings.height;
-        displaySettings.offsetX = abs((int)w - (int)displaySettings.width) / 2;
+        displaySettings.width   = (int)(aspect * displaySettings.height);
+        displaySettings.offsetX = (int)(abs((int)w - (int)displaySettings.width) / 2);
         if (displaySettings.width > (float)w) {
             displaySettings.offsetX = 0;
             displaySettings.width   = w;
